+++ src/pages/api/trpc/[trpc].ts
@@ -1,7 +1,8 @@
 import * as trpcNext from "@trpc/server/adapters/next";
-import { appRouter } from "../../../server/trpc";
+import { createContext } from "../../../server/trpc/context";
+import { appRouter } from "../../../server/trpc/_app";
 
 export default trpcNext.createNextApiHandler({
   router: appRouter,
-  createContext: () => ({}),
+  createContext,
 });
