+++ src/server/trpc/_app.ts
@@ -1,0 +1,8 @@
+import { router } from ".";
+import { transactionRouter } from "./transactions";
+
+export const appRouter = router({
+  transactions: transactionRouter,
+});
+
+export type AppRouter = typeof appRouter;
