+++ src/server/trpc/index.ts
@@ -1,12 +1,7 @@
 import { initTRPC } from "@trpc/server";
+import { Context } from "./context";
 
-const t = initTRPC.create();
+const t = initTRPC.context<Context>().create();
 
 export const router = t.router;
 export const procedure = t.procedure;
-
-export const appRouter = router({
-  // Add your routes here
-});
-
-export type AppRouter = typeof appRouter;
